requirements.txt
Django==3.0.3
pytest-django==3.8.0
python-decouple==3.3
coverage==5.0.3
psycopg2-binary==2.8.5
